user_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==19.13.0 
email-validator==2.0.0
PyJWT==2.10.0
